+++ test/e2e/helpers_test.go
@@ -6,7 +6,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"k8s.io/apimachinery/pkg/api/resource"
 	"log"
 	"math"
 	"net/http"
@@ -16,12 +15,15 @@ import (
 	"strings"
 	. "time"
 
+	"k8s.io/apimachinery/pkg/api/resource"
+
 	hzClient "github.com/hazelcast/hazelcast-go-client"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"gopkg.in/yaml.v3"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
@@ -38,9 +40,9 @@ import (
 	"github.com/hazelcast/hazelcast-platform-operator/test"
 )
 
-func GetBackupSequence(t Time) string {
+func GetBackupSequence(t Time, lk types.NamespacedName) string {
 	By("Finding Backup sequence")
-	logs := InitLogs(t)
+	logs := InitLogs(t, lk)
 	scanner := bufio.NewScanner(logs)
 	test.EventuallyInLogs(scanner, 10*Second, logInterval).Should(ContainSubstring("Starting new hot backup with sequence"))
 	line := scanner.Text()
@@ -59,10 +61,10 @@ func GetBackupSequence(t Time) string {
 	return seq
 }
 
-func InitLogs(t Time) io.ReadCloser {
+func InitLogs(t Time, lk types.NamespacedName) io.ReadCloser {
 	logs := test.GetPodLogs(context.Background(), types.NamespacedName{
-		Name:      hzName + "-0",
-		Namespace: hzNamespace,
+		Name:      lk.Name + "-0",
+		Namespace: lk.Namespace,
 	}, &corev1.PodLogOptions{
 		Follow:    true,
 		SinceTime: &metav1.Time{Time: t},
@@ -74,31 +76,36 @@ func CreateHazelcastCR(hazelcast *hazelcastcomv1alpha1.Hazelcast) {
 	By("creating Hazelcast CR", func() {
 		Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
 	})
-
-	lookupKey := types.NamespacedName{Name: hazelcast.Name, Namespace: hazelcast.Namespace}
-	By("checking Hazelcast CR running", func() {
+	lk := types.NamespacedName{Name: hazelcast.Name, Namespace: hazelcast.Namespace}
+	By("Checking Hazelcast CR running", func() {
 		hz := &hazelcastcomv1alpha1.Hazelcast{}
 		Eventually(func() bool {
-			err := k8sClient.Get(context.Background(), lookupKey, hz)
+			err := k8sClient.Get(context.Background(), lk, hz)
 			Expect(err).ToNot(HaveOccurred())
 			return isHazelcastRunning(hz)
 		}, 5*Minute, interval).Should(BeTrue())
 	})
 }
 
+func CreateHazelcastCRWithoutCheck(hazelcast *hazelcastcomv1alpha1.Hazelcast) {
+	By("Creating Hazelcast CR", func() {
+		Expect(k8sClient.Create(context.Background(), hazelcast)).Should(Succeed())
+	})
+}
+
 func RemoveHazelcastCR(hazelcast *hazelcastcomv1alpha1.Hazelcast) {
 	Expect(k8sClient.Delete(context.Background(), hazelcast, client.PropagationPolicy(metav1.DeletePropagationForeground))).Should(Succeed())
 	assertDoesNotExist(types.NamespacedName{
-		Name:      hzName + "-0",
-		Namespace: hzNamespace,
+		Name:      hazelcast.Name + "-0",
+		Namespace: hazelcast.Namespace,
 	}, &corev1.Pod{})
 
 	By("waiting for Hazelcast CR to be removed", func() {
 		Eventually(func() error {
 			h := &hazelcastcomv1alpha1.Hazelcast{}
 			return k8sClient.Get(context.Background(), types.NamespacedName{
-				Name:      hzName,
-				Namespace: hzNamespace,
+				Name:      hazelcast.Name,
+				Namespace: hazelcast.Namespace,
 			}, h)
 		}, 1*Minute, interval).ShouldNot(Succeed())
 	})
@@ -114,14 +121,10 @@ func DeletePod(podName string, gracePeriod int64) {
 	}
 }
 
-func GetHzClient(ctx context.Context, unisocket bool) *hzClient.Client {
-	var lookupKey = types.NamespacedName{
-		Name:      hzName,
-		Namespace: hzNamespace,
-	}
+func GetHzClient(ctx context.Context, lk types.NamespacedName, unisocket bool) *hzClient.Client {
 	s := &corev1.Service{}
 	Eventually(func() bool {
-		err := k8sClient.Get(context.Background(), lookupKey, s)
+		err := k8sClient.Get(context.Background(), lk, s)
 		Expect(err).ToNot(HaveOccurred())
 		return len(s.Status.LoadBalancer.Ingress) > 0
 	}, 1*Minute, interval).Should(BeTrue())
@@ -153,9 +156,9 @@ func GetClientSet() *kubernetes.Clientset {
 	return clientSet
 }
 
-func FillTheMapData(ctx context.Context, unisocket bool, mapName string, mapSize int) {
+func FillTheMapData(ctx context.Context, lk types.NamespacedName, unisocket bool, mapName string, mapSize int) {
 	var m *hzClient.Map
-	clientHz := GetHzClient(ctx, unisocket)
+	clientHz := GetHzClient(ctx, lk, unisocket)
 	By("using Hazelcast client")
 	m, err := clientHz.GetMap(ctx, mapName)
 	Expect(err).ToNot(HaveOccurred())
@@ -171,8 +174,9 @@ func FillTheMapWithHugeData(ctx context.Context, mapName string, mapSizeInGb str
 	hzAddress := fmt.Sprintf("%s.%s.svc.cluster.local:%d", hzConfig.Name, hzConfig.Namespace, n.DefaultHzPort)
 	var m *hzClient.Map
 	clientPod := CreateClientPod(hzAddress, mapSizeInGb, mapName)
+	defer DeletePod(clientPod.Name, 0)
 	mapSize, _ := strconv.ParseFloat(mapSizeInGb, 64)
-	client := GetHzClient(ctx, false)
+	client := GetHzClient(ctx, types.NamespacedName{Name: hzConfig.Name, Namespace: hzConfig.Namespace}, false)
 	m, _ = client.GetMap(ctx, mapName)
 	Eventually(func() (int, error) {
 		return m.Size(ctx)
@@ -180,7 +184,6 @@ func FillTheMapWithHugeData(ctx context.Context, mapName string, mapSizeInGb str
 	// 1310.72 entries per one Go routine. Formula: 1073741824 Bytes per 1Gb  / 8192 Bytes per entry / 100 go routines
 	err := client.Shutdown(ctx)
 	Expect(err).ToNot(HaveOccurred())
-	defer DeletePod(clientPod.Name, 0)
 }
 
 func CreateClientPod(hzAddress string, mapSizeInGb string, mapName string) *corev1.Pod {
@@ -220,11 +223,11 @@ func CreateClientPod(hzAddress string, mapSizeInGb string, mapName string) *core
 	return clientPod
 }
 
-func emptyHazelcast() *hazelcastcomv1alpha1.Hazelcast {
+func emptyHazelcast(lk types.NamespacedName) *hazelcastcomv1alpha1.Hazelcast {
 	return &hazelcastcomv1alpha1.Hazelcast{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      hzName,
-			Namespace: hzNamespace,
+			Name:      lk.Name,
+			Namespace: lk.Namespace,
 		},
 	}
 }
@@ -373,6 +376,19 @@ func createHazelcastClient(ctx context.Context, h *hazelcastcomv1alpha1.Hazelcas
 	client, err := hzClient.StartNewClientWithConfig(ctx, config)
 	Expect(err).To(BeNil())
 	return client
+}
+
+func emptyManagementCenter(lk types.NamespacedName) *hazelcastcomv1alpha1.ManagementCenter {
+	return &hazelcastcomv1alpha1.ManagementCenter{
+		ObjectMeta: v1.ObjectMeta{
+			Name:      lk.Name,
+			Namespace: lk.Namespace,
+		},
+	}
+}
+
+func isManagementCenterRunning(mc *hazelcastcomv1alpha1.ManagementCenter) bool {
+	return mc.Status.Phase == "Running"
 }
 
 func assertHazelcastRestoreStatus(h *hazelcastcomv1alpha1.Hazelcast, st hazelcastcomv1alpha1.RestoreState) *hazelcastcomv1alpha1.Hazelcast {
